Bug fixed: 定義ファイルから文字列として渡された時刻を含むDateTimeが正しく変換できない問題を解消。

+++ Implem.Libraries/Utilities/Types.cs
@@ -280,7 +280,7 @@ namespace Implem.Libraries.Utilities
             var str = obj.ToString();
             return obj != null && DateTime.TryParse(str, out data)
                 ? data
-                : obj.ToLong() > 0
+                : obj.ToDecimal() > 0
                     ? DateTime.FromOADate(double.Parse(str))
                     : DateTime.FromOADate(0);
         }
